# cmd/api/docs/swagger.yaml
basePath: /v1
definitions:
...
  database.Alias:
    properties:
      address:
        example: KT1CeekjGVRc5ASmgWDc658NBExetoKNuiqz
        type: string
      alias:
        example: Contract alias
        type: string
      network:
        example: babylonnet
        type: string
      slug:
        example: contract_slug
        type: string
    type: object
...
  formatter.DiffResult:
    properties:
      added:
        type: integer
      changed:
        type: integer
      left:
        items:
          items:
            $ref: '#/definitions/formatter.Item'
          type: array
        type: array
      removed:
        type: integer
      right:
        items:
          items:
            $ref: '#/definitions/formatter.Item'
          type: array
        type: array
    type: object
  formatter.Item:
    properties:
      chunk:
        type: string
      type:
        type: integer
    type: object
...
  handlers.BigMapDiffByKeyResponse:
    properties:
      key:
        type: object
      key_hash:
        type: string
      total:
        type: integer
      values:
        items:
          $ref: '#/definitions/handlers.BigMapDiffItem'
        type: array
    type: object
  handlers.BigMapDiffItem:
    properties:
      level:
        type: integer
      timestamp:
        type: string
      value:
        type: object
    type: object
  handlers.BigMapItem:
    properties:
      key:
        type: object
      key_hash:
        type: string
      key_string:
        type: string
      level:
        type: integer
      timestamp:
        type: string
      value:
        type: object
    type: object
  handlers.BigMapResponseItem:
    properties:
      count:
        type: integer
      data:
        $ref: '#/definitions/handlers.BigMapItem'
        type: object
    type: object
  handlers.CodeDiffLeg:
    properties:
      address:
        type: string
      level:
        type: integer
      network:
        type: string
      protocol:
        type: string
    required:
    - address
    - network
    type: object
  handlers.CodeDiffResponse:
    properties:
      diff:
        $ref: '#/definitions/formatter.DiffResult'
        type: object
      left:
        $ref: '#/definitions/handlers.CodeDiffLeg'
        type: object
      right:
        $ref: '#/definitions/handlers.CodeDiffLeg'
        type: object
    type: object
  handlers.Contract:
    properties:
      address:
        type: string
      alias:
        type: string
      annotations:
        items:
          type: string
        type: array
      balance:
        type: integer
      delegate:
        type: string
      delegate_alias:
        type: string
      entrypoints:
        items:
          type: string
        type: array
      fail_strings:
        items:
          type: string
        type: array
      fingerprint:
        $ref: '#/definitions/models.Fingerprint'
        type: object
      found_by:
        type: string
      hardcoded:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      language:
        type: string
      last_action:
        $ref: '#/definitions/models.BCDTime'
        type: object
      level:
        type: integer
      manager:
        type: string
      migrations_count:
        type: integer
      network:
        type: string
      profile:
        $ref: '#/definitions/handlers.ProfileInfo'
        type: object
      project_id:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      total_withdrawn:
        type: integer
      tx_count:
        type: integer
    type: object
...
  handlers.Error:
    properties:
      message:
        example: text
        type: string
    type: object
...
  handlers.ProfileInfo:
    properties:
      subscribed:
        type: boolean
    type: object
...
  models.BCDTime:
    type: object
...
  models.Contract:
    properties:
      address:
        type: string
      alias:
        type: string
      annotations:
        items:
          type: string
        type: array
      balance:
        type: integer
      delegate:
        type: string
      delegate_alias:
        type: string
      entrypoints:
        items:
          type: string
        type: array
      fail_strings:
        items:
          type: string
        type: array
      fingerprint:
        $ref: '#/definitions/models.Fingerprint'
        type: object
      found_by:
        type: string
      hardcoded:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      language:
        type: string
      last_action:
        $ref: '#/definitions/models.BCDTime'
        type: object
      level:
        type: integer
      manager:
        type: string
      migrations_count:
        type: integer
      network:
        type: string
      project_id:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      total_withdrawn:
        type: integer
      tx_count:
        type: integer
    type: object
  models.Fingerprint:
    properties:
      code:
        type: string
      parameter:
        type: string
      storage:
        type: string
    type: object
...
host: api.better-call.dev
info:
  contact: {}
  description: This is API description for Better Call Dev service.
  license: {}
  title: Better Call Dev API
  version: "1.0"
paths:
  /contract/{network}/{address}:
    get:
      consumes:
      - application/json
      description: Get full contract info
      operationId: get-contract
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Contract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract info
      tags:
      - contract
  /contract/{network}/{address}/bigmap/{ptr}:
    get:
      consumes:
      - application/json
      description: Get contract rating
      operationId: get-contract-bigmap
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      - description: Big map pointer
        in: path
        name: ptr
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BigMapResponseItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get big map by pointer
      tags:
      - contract
  /contract/{network}/{address}/bigmap/{ptr}/{key_hash}:
    get:
      consumes:
      - application/json
      description: Get big map diffs by pointer and key hash
      operationId: get-contract-bigmap-keyhash
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      - description: Big map pointer
        in: path
        name: ptr
        required: true
        type: integer
      - description: Key hash in big map
        in: path
        name: key_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BigMapDiffByKeyResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get big map diffs by pointer and key hash
      tags:
      - contract
  /contract/{network}/{address}/code:
    get:
      consumes:
      - application/json
      description: Get contract code
      operationId: get-contract-code
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract code
      tags:
      - contract
...
  /diff:
    post:
      consumes:
      - application/json
      description: Get diff between two contracts
      operationId: get-diff
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CodeDiffResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get diff between two contracts
      tags:
      - contract
...
  /pick_random:
    get:
      consumes:
      - application/json
      description: Get random contract with 2 or more operations
      operationId: get-random-contract
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contract'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Show random contract
      tags:
      - contract
...
  /slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get contract by slug
      operationId: get-contract-by-slug
      parameters:
      - description: Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Alias'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract by slug
      tags:
      - contract
...
swagger: "2.0"